cleaned up clippy warnings and moved the dice game code to a different file
